requirements.txt
autopep8==2.0.4
pycodestyle==2.11.1
Pillow==10.3.0
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pygame==2.5.2
python-dotenv==1.0.1
requests==2.32.2
pycco==0.6.0